Added generic calculator that returns a calculator for the diff types

+++ MovieRental/PriceCalculators.cs
@@ -4,12 +4,12 @@ namespace MovieRental
 {
 	public interface IPriceCalculator
 	{
-		double Calculate(int days);
+		int Calculate(int days);
 	}
 
 	public class NormalPriceCalculator : IPriceCalculator
 	{
-		public double Calculate(int days)
+		public int Calculate(int days)
 		{
 			var total = 0;
 			total += 2;
@@ -21,7 +21,7 @@ namespace MovieRental
 
 	public class KidsPriceCalculator : IPriceCalculator
 	{
-		public double Calculate(int days)
+		public int Calculate(int days)
 		{
 			var total = 0;
 			total += 3;
@@ -33,7 +33,7 @@ namespace MovieRental
 
 	public class PremierePriceCalculator : IPriceCalculator
 	{
-		public double Calculate(int days)
+		public int Calculate(int days)
 		{
 			var total = 3;
 			return total;
